requirements.txt
pandas
matplotlib
seaborn
mysql-connector-python
scikit-learn
python-docx
numpy
sqlalchemy